% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{abmi_climate}
\alias{abmi_climate}
\title{Provide users with bioclimatic lookup table}
\usage{
abmi_climate()
}
\value{
A dataframe with all of the coefficients used in the climate model. Each taxonomic group uses a unique set of these variables.
}
\description{
This function provides the lookup table of the bioclimatic covariates used in the climate model.
}
\examples{
\dontrun{
climate.lookup <- abmi_climate()
}

}
