% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{abmi_landcover}
\alias{abmi_landcover}
\title{Internal lookup tables}
\usage{
abmi_landcover(landcover)
}
\arguments{
\item{landcover}{Define 'Vegetation' or 'Soil' landcover classes.}
}
\value{
A dataframe with all of the coefficients used in the joint climate and landcover model.
}
\description{
This function provides the lookup table of the vegetation, soil, and footprint definitions
}
\section{Provide users with access to internal lookup tables.}{
NA
}

\examples{
\dontrun{
landcover.lookup <- abmi_landcover(landcover = "Vegetation")
}

}
