% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{abmi_species}
\alias{abmi_species}
\title{Provide users with species lookup table}
\usage{
abmi_species()
}
\value{
A dataframe with all of the species information include model region and validation statistics.
}
\description{
This function provides the lookup table of all species models.
}
\examples{
\dontrun{
species.lookup <- abmi_climate()
}

}
