% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioclimatic.lookup}
\alias{bioclimatic.lookup}
\title{bioclimatic-lookup}
\format{
An object of class \code{data.frame} with 20 rows and 2 columns.
}
\usage{
bioclimatic.lookup
}
\description{
bioclimatic-lookup
}
\keyword{datasets}
