% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abmi_load_bioclimatic.R
\name{bioclimatic}
\alias{bioclimatic}
\alias{abmi_load_bioclimatic}
\title{Bioclimatic data}
\usage{
abmi_load_bioclimatic(dir = NULL)
}
\arguments{
\item{dir}{Directory path.}
}
\value{
`abmi_load_bioclimatic Loads the bioclimatic data into the current session.
}
\description{
Load the bioclimatic raster into memory
}
