% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend_predict.R
\name{blend_predict}
\alias{blend_predict}
\title{Blend species predictions}
\usage{
blend_predict(climate = NULL, veg = NULL, soil = NULL)
}
\arguments{
\item{climate}{Matrix of the bioclimatic variables that is matched to the vegetation and/or soil matrices.}

\item{veg}{Vector of vegetation based predictions generated by the species_predict function.}

\item{soil}{Vector of soil based predictions generated by the species_predict function.}
}
\description{
Function for merging the vegetation and soil based predictions.
}
