% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_adjustment.R
\name{coefficient_adjustment}
\alias{coefficient_adjustment}
\title{Coefficient Adjustment}
\usage{
coefficient_adjustment(model, coef = NULL, value = NULL)
}
\arguments{
\item{model}{Vector defined as "Vegetation", "Soil", or "All" so the appropriate model is adjusted.}

\item{coef}{Vector of characters that the user wishes to modify.}

\item{value}{User defined value to be assigned to the specified coefficients. Adjustment is made across all bootstraps.}
}
\description{
This function will adjust the stored coefficients to user defined values. 
In general, we advice not adjusting these coefficients. However, there are specific situations 
where adjusting stored coefficients to 0 (e.g., Crop, Urban Industrial, Hard Linear) may be appropriate.
}
