% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_extraction.R
\name{coefficient_extraction}
\alias{coefficient_extraction}
\title{Coefficient extraction}
\usage{
coefficient_extraction(species, model = NULL)
}
\arguments{
\item{species}{Unique Species ID. If defined as "All", the complete coefficient table will be provided to the user.}

\item{model}{Vector defined as "Vegetation" or "Soil" that provides the relevant coefficients.}
}
\description{
This function extracts the stored coefficients and provides a table or summary figure.
}
