% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{coefs}
\alias{coefs}
\alias{abmi_dowload_coefs}
\alias{abmi_load_coefs}
\alias{abmi_unload_coefs}
\title{Coefficients}
\usage{
abmi_dowload_coefs(dir = NULL, ...)

abmi_load_coefs(dir = NULL)

abmi_unload_coefs()
}
\arguments{
\item{dir}{Directory path.}

\item{...}{Arguments parameters passed to [utils::download.file()].}
}
\value{
`abmi_dowload_coefs` dowloads a file as a side effect.

`abmi_load_coefs` loads the coefficients into the current session.

`abmi_unload_coefs` unloads the coefficients from the current session.
}
\description{
Download, load, unload coefficients.
}
\seealso{
[utils::download.file()].
}
