% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_climate.R
\name{.create_climate}
\alias{.create_climate}
\title{Create bioclimatic data}
\usage{
.create_climate(climate.grid, taxon)
}
\arguments{
\item{climate.grid}{Matrix containing the stored bioclimatic variables. Rownames are used as the cell identifier.}

\item{taxon}{Taxonomic group associated with the focal species of interest.}
}
\description{
Function for generating the taxon specific bioclimatic data.
}
