% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxon.R
\name{.get_taxon}
\alias{.get_taxon}
\title{Determine taxonomic group}
\usage{
.get_taxon(species)
}
\arguments{
\item{species}{Valid species name available from the abmi_species function.}

\item{taxon}{Taxonomic group associated with the focal species of interest.}
}
\description{
Function for determining a species taxonomic group
}
