% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mammal_predict.R
\name{.mammal_predicts}
\alias{.mammal_predicts}
\title{Mammal predictions}
\usage{
.mammal_predicts(
  species,
  taxon = NULL,
  veg = NULL,
  soil = NULL,
  climate = NULL,
  boot = 0,
  modified = FALSE
)
}
\arguments{
\item{species}{Unique Species ID.}

\item{taxon}{Taxon associated with the unique Species ID.}

\item{veg}{Matrix of standardized vegetation information (row sum equals 1). If NULL and model is available, no prediction is generated.}

\item{soil}{Matrix of standardized soil information (row sum equals 1). If NULL and model is available, no prediction is generated.}

\item{climate}{Matrix of the bioclimatic variables that is matched to the vegetation and/or soil matrices.}

\item{boot}{Defines bootstrap iteration to generate the predictions from.}
}
\description{
Function for calculating the species level predictions for mammals
}
