% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_climate.R
\name{extract_climate}
\alias{extract_climate}
\title{Extract climate}
\usage{
extract_climate(spatial.grid, cell.id = NULL, reproject = FALSE)
}
\arguments{
\item{spatial.grid}{Terra spatial vector of points that the user wants to extract bioclimatic information for}

\item{cell.id}{Vector of cell ID used to link the bioclimatic and landcover data.}

\item{reproject}{Logical; If TRUE, will call terra::project to reproject to the appropriate CRS (NAD83 / Alberta 10-TM (Forest) (EPSG:3400))}
}
\description{
Function for extracting the relevant bioclimatic variables
}
