% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_unknown_ages.R
\name{fill_unknown_ages}
\alias{fill_unknown_ages}
\title{Fill Unknown Ages}
\usage{
fill_unknown_ages(land.cover, age.distribution = "age.old.nsr.maltman")
}
\arguments{
\item{land.cover}{The dataframe with unknown ages that needs to be updated.}

\item{age.distribution}{The data file that contains average age distributions by natural subregion.}
}
\value{
A data frame with missing ages filled in.
}
\description{
There are often parts of forested landcover datasets where the age of the
forest is unknown. When this occurs, these areas must be assigned an age.
This function assigns ages to these areas based on the average forest age
distributions of Natural Subregions in Alberta. These distributions have been
shifted such that the minimum age is 80 years, assuming that the areas of
unknown age have not been harvested or burned, and therefore are likely to be
at least 80 years old.
}
\details{
Note that the dataframe should have all the necessary columns, in order. If
the columns are not found, an error will be returned. If they are out of 
order,the function will try to put them back in the correct order. The
correct order for each veg type follows this pattern: 
PineR, Pine1, ..., Pine8.
}
