% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inv_link_function}
\alias{inv_link_function}
\title{model-inverse-link-functions}
\format{
An object of class \code{list} of length 7.
}
\usage{
inv_link_function
}
\description{
model-inverse-link-functions
}
\keyword{datasets}
