% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landcover.names}
\alias{landcover.names}
\title{landcover-lookup}
\format{
An object of class \code{list} of length 2.
}
\usage{
landcover.names
}
\description{
landcover-lookup
}
\keyword{datasets}
