% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{link_function}
\alias{link_function}
\title{model-link-functions}
\format{
An object of class \code{list} of length 7.
}
\usage{
link_function
}
\description{
model-link-functions
}
\keyword{datasets}
