% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate.R
\name{plot_climate}
\alias{plot_climate}
\title{Plot climate}
\usage{
plot_climate(variable = NULL)
}
\arguments{
\item{variable}{Viable covariate from the abmi_climate() lookup table (Column)}
}
\description{
Function for visualizing the relevant bioclimatic variables
}
