% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coef.R
\name{plot_coef}
\alias{plot_coef}
\title{Plot species coefficient}
\usage{
plot_coef(species, model = NULL)
}
\arguments{
\item{species}{Unique Species ID.}

\item{model}{Vector defined as "vegetation" or "soil" that provides the relevant coefficients.}
}
\description{
Function for visualizing the species coefficients
}
