% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species.R
\name{plot_species}
\alias{plot_species}
\title{Plot species}
\usage{
plot_species(spat.raster = NULL, variable = NULL)
}
\arguments{
\item{spat.raster}{SpatRaster object of the predictions.}

\item{variable}{Attribute within the SpatRaster object for plotting.}
}
\description{
Function for visualizing the predictions from species_predicts
}
