% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_predict.R
\name{species_predict}
\alias{species_predict}
\title{Species predictions}
\usage{
species_predict(
  species,
  veg = NULL,
  soil = NULL,
  climate = NULL,
  modified = FALSE,
  boot = 0
)
}
\arguments{
\item{species}{Unique Species ID.}

\item{veg}{Matrix of standardized vegetation information (row sum equals 1). If NULL and model is available, no prediction is generated.}

\item{soil}{Matrix of standardized soil information (row sum equals 1). If NULL and model is available, no prediction is generated.}

\item{climate}{Matrix of the bioclimatic variables that is matched to the vegetation and/or soil matrices.}

\item{modified}{Logical; Default behaviour (FALSE) means the original species coefficients are used. TRUE indicators the function should use the coefficients created through the coefficient_adjustment function.}

\item{boot}{Defines bootstrap iteration to generate the predictions from. If 0, uses the median bootstrap run unique for each species.}
}
\description{
Wrapper function for calculating the species level predictions
}
