% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_predict.R
\name{truncate_predict}
\alias{truncate_predict}
\title{Truncate prediction}
\usage{
truncate_predict(species, current = NULL, reference = NULL)
}
\arguments{
\item{species}{Unique Species ID.}

\item{current}{Blended or single model type predictions (e.g., vegetation or soil) based on current landscape conditions}

\item{reference}{Blended or single model type predictions (e.g., vegetation or soil) based on reference landscape conditions}
}
\description{
Function for truncating the predictions from the species_predicts and blend_predict function. Users can use truncate based on the current landscape predictions (current) or truncate two time periods using the same threshold.
}
