% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABMIexploreR.R
\docType{package}
\name{ABMIexploreR}
\alias{ABMIexploreR-package}
\alias{ABMIexploreR}
\title{ABMIexploreR}
\description{
Decision support tool that provides an interface to enable custom predictions based on species distribution models created by the Alberta Biodiversity Monitoring Institute (ABMI) in collaboration with the Boreal Avian Modelling (BAM) Project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/beallen/ABMIexploreR}
  \item Report bugs at \url{https://github.com/beallen/ABMIexploreR/issues}
}

}
\author{
Brandon Allen \email{brandon.allen@ualberta.ca}
}
